requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
rapidfuzz==3.6.0
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.0.1
